resume-analyzer/requirements.txt
streamlit
nltk==3.6.3
spacy==3.1.3
pandas
pyresparser
pdfminer.six
streamlit-tags
Pillow
PyMySQL==1.0.2
pafy
plotly
youtube-dl
Flask==2.0.1
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0